requirements-deploy.txt
aws-cdk-lib==2.62.0
constructs>=10.0.0,<11.0.0
tomli==2.0.1